Added mdbootstrap, Added Footer, Organized Files
