+++ ClientApp/package.json
@@ -3,10 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.4.0",
     "axios": "^1.4.0",
     "bootstrap": "^5.2.3",
     "http-proxy-middleware": "^2.0.6",
     "jquery": "^3.6.4",
+    "mdb-react-ui-kit": "^6.1.0",
     "merge": "^2.1.1",
     "oidc-client": "^1.11.5",
     "react": "^18.2.0",
