+++ ClientApp/src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import AppRoutes from './AppRoutes';
 import AuthorizeRoute from './components/api-authorization/AuthorizeRoute';
-import { Layout } from './components/Layout';
+import { Layout } from './components/layout/Layout';
 import './custom.css';
 
 export default class App extends Component {
