+++ ClientApp/src/components/function/functions.php
