+++ ClientApp/src/components/layout/Layout.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { Container } from 'reactstrap';
-import { NavMenu } from './NavMenu';
+import { NavMenu } from './navbar/NavMenu';
+import { Footer } from './footer/Footer';
 
 export class Layout extends Component {
   static displayName = Layout.name;
@@ -12,6 +13,7 @@ export class Layout extends Component {
         <Container tag="main">
           {this.props.children}
         </Container>
+        <Footer />
       </div>
     );
   }
