+++ ClientApp/src/components/layout/footer/Footer.js
@@ -1,0 +1,13 @@
+import React,{ Component } from "react";
+import {  } from 'reactstrap';
+import './Footer.css';
+
+export class Footer extends Component{
+    render() {
+        return (
+            <footer>
+                <p>Practice Website @ 2023</p>
+            </footer>
+        );
+    }
+}
