+++ ClientApp/src/components/layout/navbar/NavMenu.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Collapse, Navbar, NavbarBrand, NavbarToggler, NavItem, NavLink } from 'reactstrap';
 import { Link } from 'react-router-dom';
-import { LoginMenu } from './api-authorization/LoginMenu';
+import { LoginMenu } from '../../api-authorization/LoginMenu';
 import './NavMenu.css';
 
 export class NavMenu extends Component {
